Move shared functionality to abstract class
